% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.hci.R
\name{confint.hci}
\alias{confint.hci}
\title{Confidence intervals for \code{hci} objects}
\usage{
\method{confint}{hci}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{hci}}

\item{parm}{Unused}

\item{level}{Confidence interval level defaults to \code{0.95}}

\item{...}{Unused}
}
\value{
A confidence interval in a numeric vector of length 2
}
\description{
Confidence intervals for \code{hci} objects
}
\examples{
data(housing)
ci.bmi <- ci(ineqvar = housing$income, outcome = housing$bmi, method = "direct")
confint(ci.bmi)

}
