% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected-value.R
\name{corrected_value}
\alias{corrected_value}
\title{Corrected value}
\usage{
corrected_value(x)
}
\arguments{
\item{x}{Object to return corrected value of}
}
\value{
Returns the corrected value stored in the object
}
\description{
Corrected value
}
