% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decomposition}
\alias{decomposition}
\title{Decomposition analysis}
\usage{
decomposition(outcome, betas, mm, ranker, wt, correction, citype = "CI")
}
\arguments{
\item{outcome}{Outcome variable}

\item{betas}{Beta coefficients from regression.}

\item{mm}{Model matrix from regression}

\item{ranker}{Ranking variable}

\item{wt}{Weights}

\item{correction}{Apply sign correction?}

\item{citype}{CI type to be calculated, defaults to \code{CI}. Use \code{CIw} for binary outcomes.}
}
\value{
S3 object of class decomposition
}
\description{
Decomposition analysis
}
