% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{Artificial example data on housing conditions}
\format{
\code{data.frame} object.

Variable list:
\describe{
\item{id}{unique identifier per person}
\item{sex}{female or male}
\item{age}{integer, from 20 to 94}
\item{tenure}{categorical. One of homeless, irregular, own_apartment, own_house, or rent}
\item{height}{height in cm }
\item{weight}{weight in kg}
\item{bmi}{weight/(height/100)^2}
\item{income}{continuous, imaginary currency without unit}
}
}
\source{
Artificially generated by the package authors
}
\usage{
data(housing)
}
\description{
Microdata with a permissive license that includes continuous data on health and income is hard to come by. In stead of real data, the package thus includes an imaginary dataset.
}
\keyword{datasets}
