% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hci.R
\name{plot.hci}
\alias{plot.hci}
\title{Plots the concentration curve for an \code{hci} object.}
\usage{
\method{plot}{hci}(x, ...)
}
\arguments{
\item{x}{Object with of \code{hci}}

\item{...}{Further arguments passed to \code{base::plot()}}
}
\value{
Invisibly returns \code{x} as the function is called for side effects (plotting).
}
\description{
Plots the concentration curve for an \code{hci} object.
}
\examples{
data(housing)
ci.bmi <- ci(ineqvar = housing$income, outcome = housing$bmi, method = "direct")
plot(ci.bmi)
 
}
