% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hci.R
\name{print.hci}
\alias{print.hci}
\title{Prints an \code{hci} object.}
\usage{
\method{print}{hci}(x, ...)
}
\arguments{
\item{x}{Object of type \code{hci}}

\item{...}{Currently unused}
}
\value{
Invisibly returns \code{x} as the function is called for side effects.
}
\description{
Prints an \code{hci} object.
}
\examples{
data(housing)
ci.bmi <- ci(ineqvar = housing$income, outcome = housing$bmi, method = "direct")
print(ci.bmi)

}
