% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank-wt.R
\name{rank_wt}
\alias{rank_wt}
\title{Calculates the weighted rank}
\usage{
rank_wt(x, wt)
}
\arguments{
\item{x}{numeric vector}

\item{wt}{weights}
}
\value{
A numeric vector containing weighted fractional ranks of the elements of \code{x}.
}
\description{
Calculates the weighted rank
}
\examples{
x <- sample(1:10, size = 10, replace = TRUE)
x.weight <- seq(0, 1, length.out = 10)
rank_wt(x, wt = x.weight)
}
\references{
Kakwani \emph{et al}., 1997.
}
\author{
Peter Konings
}
