% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.decomposition.R
\name{summary.decomposition}
\alias{summary.decomposition}
\title{Summary for decomposition}
\usage{
\method{summary}{decomposition}(object, digits = getOption("digits"), addcoefs = FALSE, ...)
}
\arguments{
\item{object}{Result of a decomposition}

\item{digits}{Number of digits, defaults to R `digits` option}

\item{addcoefs}{Whether or not to add coefficients (defaults to `FALSE`)}

\item{...}{Additional parameters, currently unused}
}
\description{
Summary for decomposition
}
