% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hci.R
\name{summary.hci}
\alias{summary.hci}
\title{Prints the a summary of the concentration index object \code{hci}}
\usage{
\method{summary}{hci}(object, ...)
}
\arguments{
\item{object}{Object of type \code{hci}}

\item{...}{Currently unused}
}
\value{
No returns value. Directly prints to the standard output connection.
}
\description{
Prints the a summary of the concentration index object \code{hci}
}
\examples{
data(housing)
ci.bmi <- ci(ineqvar = housing$income, outcome = housing$bmi, method = "direct")
summary(ci.bmi)

}
