% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-sign.R
\name{correct_sign}
\alias{correct_sign}
\title{Corrects negative values in the health variable}
\usage{
correct_sign(x, shift = TRUE)
}
\arguments{
\item{x}{A numeric vector, typically representing health.}

\item{shift}{If \code{FALSE} (the default), 0 is imputed for all negative values in \code{x}. If \code{TRUE} the minimum value of \code{x} is subtracted from it.}
}
\value{
Returns a list with 2 components:
\itemize{
\item{corrected}{corrected version of \code{x}}
\item{modified}{logical, \code{TRUE} when any of the elements of \code{x} have been changed}
}
These components can be extracted with the functions \code{correctedValue} and \code{isCorrected}.
}
\description{
The Relative Concentration Index is not bonded between [-1,1] if the health variable contains both negative and positive values. This function corrects for this either by imputing a value of 0 for all negative values or by subtracting the minimum value.
}
\examples{
data(housing)

# standardize & normalize bmi, will introduce negative values
housing$zbmi <- (housing$bmi - mean(housing$bmi))/ sd(housing$bmi)

housing$zbmi.shifted <- corrected_value(correct_sign(housing$zbmi, shift = TRUE))
housing$zbmi.imputed <- corrected_value(correct_sign(housing$zbmi, shift = FALSE))

## compare the effect of both methods
plot(density(housing$zbmi, na.rm = TRUE))
points(density(housing$zbmi.shifted, na.rm = TRUE), col = 'blue')
points(density(housing$zbmi.imputed, na.rm = TRUE), col = 'green')

}
\author{
Peter Konings
}
