% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.decomposition.R
\name{plot.decomposition}
\alias{plot.decomposition}
\title{Plots a barplot of the contribution percentages. Sets custom plot margins and uses the graphical parameters `xlim`, `horiz`, `las` and `xlab` which therefore cannot be customized}
\usage{
\method{plot}{decomposition}(x, decreasing = TRUE, horiz = F, ...)
}
\arguments{
\item{x}{Object returned from decomposition function}

\item{decreasing}{Whether to sort contributions decreasing or not}

\item{horiz}{If the barplots should be printed horizontally or vertically}

\item{...}{Graphical parameter passed on to `barplot`}
}
\description{
Plots a barplot of the contribution percentages. Sets custom plot margins and uses the graphical parameters `xlim`, `horiz`, `las` and `xlab` which therefore cannot be customized
}
