% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.decomposition.R
\name{print.decomposition}
\alias{print.decomposition}
\title{Print function for \code{decomposition} objects.}
\usage{
\method{print}{decomposition}(x, ...)
}
\arguments{
\item{x}{Object of type \code{decomposition}}

\item{...}{Currently unused}
}
\value{
Invisibly returns \code{x} as the function is called for side effects.
}
\description{
Print function for \code{decomposition} objects.
}
\examples{
data(housing)
# Linear regression &  decompose 
fit.lm <- lm(bmi ~ sex + tenure + place + age,data = housing)
contrib.lm <- contribution(fit.lm, housing$income)

# print
print(contrib.lm)
}
