\name{rpart_ci}
\alias{rpart_ci}
\title{Recursive Partitioning and Regression Trees Using the Concentration Index}

\description{
TO DO
}

\usage{
rpart_ci(formula, data, weights, type = c("CI", "CIg", "CIc"),
  subset, na.action = na.rpart, model = FALSE, x = FALSE, y = TRUE,
  control, cost, ...)
}

\arguments{
  \item{formula}{}
  \item{data}{}
  \item{weights}{}
  \item{type}{}
  \item{subset}{}
  \item{na.action}{}
  \item{model}{}
  \item{x}{}
  \item{y}{}
  \item{control}{}
  \item{cost}{}
  \item{\dots}{}
}

\value{
An object of class \code{rpart}. See \code{\link{rpart.object}}.
}

\references{
#TO DO
}

\seealso{
#TO DO
}

\examples{
#TO DO
}